+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Added basic support for yaw control in drone shows, including:
+  - multiple options for selecting a drone template before creating the swarm and the takeoff grid: sphere, cone (suitable for yaw control) or any custom selected object;
+  - yaw angle export into the .skyc show format.
+
+### Fixed
+
+- Minimum navigation altitude during export fixed for shows with staged landing
+
 ## [3.0.2] - 2023-11-20
 
 ### Fixed